[Tool] 아이템 생성툴에 골드, 실버 재화 추가, 삭제 기능 추가

+++ Scripts/SaveData/PlayerData.cs
@@ -228,7 +228,7 @@ namespace GGemCo.Scripts
         /// <param name="currencyType"></param>
         /// <param name="value"></param>
         /// <returns></returns>
-        public ResultCommon MinusCurrency(CurrencyConstants.Type currencyType, int value)
+        public ResultCommon MinusCurrency(CurrencyConstants.Type currencyType, long value)
         {
             switch (currencyType)
             {
